scripts/requirements.txt
python-docx==0.8.11
markdown==3.4.4 